eval-annotations: tolerate empty string defined annotations
It is not the responsibility of the function eval_annotations() to
reject an empty string annotation, but rather the responsibility of
the callback function to be called with that annotation as an
argument.

For example, one might specify an repo project annotation
VERIFY_0_REFS_EXCLUDE with an empty content. This case is legitimate
and means that no refs should be excluded from the refs list matching
the VERIFY_0_REFS glob patterns.
In that very case, eval_annotations() should not reject the annotation
based on the fact that it is empty.
The callback function verify() manages then properly the fact that the
argument filled with the content of VERIFY_0_REFS_EXCLUDE is empty.

Notice that it is also possible to specify no VERIFY_0_REFS_EXCLUDE
annotation for the same result (the callback function verify() is
variadic), hence the tricky parsing of the annotations in the next
commit.
The callback function verify() manages also properly the fact that the
argument with the content of VERIFY_0_REFS_EXCLUDE is missing.

Change-Id: I457c9e9301bda819d637ab5b4be07c70573acda7