yield-branches-from-remote: new "repo forall" convenience script
This script "yields" (see the last paragraph for the rationale behind
this verb) all the remote branches (the remote here is the default
remote, it does not include the upstream remotes potentially defined
in the UPSTREAM annotations) as local branches. These local branches
properly tracks the appropriate branches on the default remote.

This script is needed because of "repo sync" creating "detached HEAD"
repositories by default. As a consequence, the project repositories
found in the repo source tree do not hold any local branches. This is
an issue when the user wants to verify the OpenPGP signatures of the
branch heads commits with the upcoming "verify.sh" script. Since there
is absolutely no branches in the repositories, the verify script do
not match any branch references and exits successfully without having
verified anything. Such a script was then needed to populate
automatically all the repositories with local branches.

I did not want the verb "fetch" (which would have been appropriate and
intuitive) in order not to confuse the developer since the "fetch"
verb has a precise meaning in the Git nomenclature.
However this script does absolutely zero "git fetch" and may be used
offline (provided that all the refs are still valid in the git
objects).
If the user still wants to fetch the remote branches and yield them
locally with this script, here is a short command template that does
this (provided that the user has activated the CLIP OS toolkit for the
PATH resolution of the "yield-branches-from-remote.sh" script):

  $ repo forall [<project>...] -c 'git fetch && yield-branches-from-remote.sh'

Change-Id: Ie17c484a6ab4ddcffd759534b74b196360166b75