source_me.sh: improve notice message and the Git LFS repositories check
The previous version was checking the proper activation of the Git LFS
filters in the Git configuration of the user environment. It was not
checking the fact that all the Git LFS objects have been properly
checked out in the repositories backed by Git LFS. This triggered
wrongfully errors in the case where the source tree were not fetched
and synchronized in the same environment as the one running it
(typically in the containeraized CI environemnts).

Change-Id: I3eba394f7e84373260c0787439880badaf29f17c