Python deps: get rid of Safety (pyup.io tool)
Getting rid of Safety tool (CLI tool to check the "requirements.txt"
file for commonly known vulnerable Python package dependencies) as
this is not really useful since we already host our source code on
GitHub which already provides such warnings.

Also, Safety provides more warnings when used with a paid API key
which was not the case for us. As a consequence, the GitHub security
alerts are more thorough than running manually the safety tool.

Incidentally, this enables us to drop the transitive dependency to
PyYAML 3.13 (which is marked as vulnerable to CVE-2017-18342 by GitHub
security alerts).

Change-Id: Ic30cf4a285173926f51b2d649350f7e5a8d65e97
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>