CLIP OS toolkit: Bump Python package dependencies
This Python packages bump is required in order to upgrade Pip to
version 19.1.1 which fixes a bug causing "pyproject.toml" file to be
considered eventhough "--no-use-pep517" is specified on the command
line (see https://github.com/pypa/pip/issues/6434 ).
This bug was causing our CI systems to fail on the install of package
sphinx-autodoc-typehints which makes use of "pyproject.toml" file.

Change-Id: Ie102facaf16baac7699e92237db61c3ccb817783
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>